refactor: move dev_commands under dev
