+++ deployer/dev/commands/__init__.py
@@ -1,0 +1,26 @@
+import deployer.dev.commands.cilogon  # noqa: F401
+import deployer.dev.commands.config.get_clusters  # noqa: F401
+import deployer.dev.commands.debug  # noqa: F401
+import deployer.dev.commands.develop.use_cluster_credentials  # noqa: F401
+import deployer.dev.commands.exec.aws.aws_app  # noqa: F401
+import deployer.dev.commands.exec.azure  # noqa: F401
+import deployer.dev.commands.exec.cost_monitoring.cost_monitoring_app  # noqa: F401
+import deployer.dev.commands.exec.get_quota_usage  # noqa: F401
+import deployer.dev.commands.exec.infra_components  # noqa: F401
+import deployer.dev.commands.exec.promql  # noqa: F401
+import deployer.dev.commands.generate.cryptnono_config  # noqa: F401
+import deployer.dev.commands.generate.dedicated_cluster.aws  # noqa: F401
+import deployer.dev.commands.generate.dedicated_cluster.gcp  # noqa: F401
+import deployer.dev.commands.generate.hub_asset.cluster_entry  # noqa: F401
+import deployer.dev.commands.generate.hub_asset.hub_files  # noqa: F401
+import deployer.dev.commands.generate.mau  # noqa: F401
+import deployer.dev.commands.generate.placeholders  # noqa: F401
+import deployer.dev.commands.generate.resource_allocation.daemonset_requests  # noqa: F401
+import deployer.dev.commands.generate.resource_allocation.generate_choices  # noqa: F401
+import deployer.dev.commands.generate.resource_allocation.instance_capacities  # noqa: F401
+import deployer.dev.commands.generate.resource_allocation.update_nodeinfo  # noqa: F401
+import deployer.dev.commands.grafana.central_grafana  # noqa: F401
+import deployer.dev.commands.grafana.tokens  # noqa: F401
+import deployer.dev.commands.keys.decrypt_age  # noqa: F401
+import deployer.dev.commands.transform.cost_table  # noqa: F401
+import deployer.dev.commands.update.eksctl  # noqa: F401
