+++ deployer/dev/commands/exec/cost_monitoring/cost_monitoring_app.py
@@ -11,7 +11,7 @@ from yarl import URL
 
 from deployer.app import exec_app
 from deployer.commands.validate.config import cluster_config as validate_cluster_config
-from deployer.dev_commands.exec.aws.aws_app import setup_aws_sts_env
+from deployer.dev.commands.exec.aws.aws_app import setup_aws_sts_env
 from deployer.infra_components.cluster import Cluster
 
 cost_monitoring = typer.Typer(pretty_exceptions_show_locals=False)
