+++ deployer/dev/commands/generate/cryptnono_config/__init__.py
