+++ deployer/dev/commands/generate/dedicated_cluster/dedicated_cluster_app.py
