+++ deployer/dev/commands/generate/dedicated_cluster/gcp.py
