+++ deployer/dev/commands/generate/hub_asset/cluster_entry.py
