+++ deployer/dev/commands/generate/hub_asset/hub_asset_app.py
