+++ deployer/dev/commands/generate/resource_allocation/daemonset_requests.yaml
