+++ deployer/dev/commands/generate/resource_allocation/generate_choices.py
