+++ deployer/dev/commands/generate/resource_allocation/instance_capacities.yaml
