+++ deployer/dev/commands/generate/resource_allocation/node-capacity-info.json
