+++ deployer/dev/commands/generate/resource_allocation/resource_allocation_app.py
