+++ deployer/dev/commands/generate/resource_allocation/update_nodeinfo.py
