+++ deployer/dev/commands/update/eksctl.py
@@ -1,7 +1,7 @@
 import typer
 
 from deployer.app import update_app
-from deployer.dev_commands.generate.dedicated_cluster.aws import generate_eksctl
+from deployer.dev.commands.generate.dedicated_cluster.aws import generate_eksctl
 from deployer.infra_components.cluster import Cluster
 from deployer.utils.rendering import print_colour
