--- deployer/dev_commands/__init__.py
@@ -1,26 +1,0 @@
-import deployer.dev_commands.cilogon  # noqa: F401
-import deployer.dev_commands.config.get_clusters  # noqa: F401
-import deployer.dev_commands.debug  # noqa: F401
-import deployer.dev_commands.develop.use_cluster_credentials  # noqa: F401
-import deployer.dev_commands.exec.aws.aws_app  # noqa: F401
-import deployer.dev_commands.exec.azure  # noqa: F401
-import deployer.dev_commands.exec.cost_monitoring.cost_monitoring_app  # noqa: F401
-import deployer.dev_commands.exec.get_quota_usage  # noqa: F401
-import deployer.dev_commands.exec.infra_components  # noqa: F401
-import deployer.dev_commands.exec.promql  # noqa: F401
-import deployer.dev_commands.generate.cryptnono_config  # noqa: F401
-import deployer.dev_commands.generate.dedicated_cluster.aws  # noqa: F401
-import deployer.dev_commands.generate.dedicated_cluster.gcp  # noqa: F401
-import deployer.dev_commands.generate.hub_asset.cluster_entry  # noqa: F401
-import deployer.dev_commands.generate.hub_asset.hub_files  # noqa: F401
-import deployer.dev_commands.generate.mau  # noqa: F401
-import deployer.dev_commands.generate.placeholders  # noqa: F401
-import deployer.dev_commands.generate.resource_allocation.daemonset_requests  # noqa: F401
-import deployer.dev_commands.generate.resource_allocation.generate_choices  # noqa: F401
-import deployer.dev_commands.generate.resource_allocation.instance_capacities  # noqa: F401
-import deployer.dev_commands.generate.resource_allocation.update_nodeinfo  # noqa: F401
-import deployer.dev_commands.grafana.central_grafana  # noqa: F401
-import deployer.dev_commands.grafana.tokens  # noqa: F401
-import deployer.dev_commands.keys.decrypt_age  # noqa: F401
-import deployer.dev_commands.transform.cost_table  # noqa: F401
-import deployer.dev_commands.update.eksctl  # noqa: F401
