--- deployer/dev_commands/generate/cryptnono_config/enc-blocklist-generator.secret.py
@@ -1,21 +1,0 @@
-{
-	"data": "ENC[AES256_GCM,data:/Y493SE9/q667E/OULeDdb4CSw8IIcuXCIM13lH83yNzaejP+XOJRa73ROSshvPaduFUsiWNgZj8PDLqgpspGolzGGFV4FFGrSRtEMS60RPT7pbmbBSXtUqC4hgyfTxICF15nzZcbbPz0wFVe6Ed7Z3HF/BdFc+1DjTvRhAnWXLTPmrlnt975KaZF7XaIyWHMogqcnKwku7az5eFVpNJhTylyMUQb/Tg/kncwFHSSpF/UIsEpZ2VoikcRUhOw3Qk44y5kpDqUxTqGEWtgrd3q2mo2xbCubNaUwPcKECpqukmky9LQlPFcfVeB0Lr2s4o0MDrYfhw9tl7cJsMObQdoggTgLM6/IA4rj70O6TrtIJKS7ZC7Q2ZWjebhUhHA9TAnCUNgmqVeJgNL928FPnTVbhRqUr5Enj+yKiptz4eBzMMHalw9prVo/D1LyNu8Q7F+IotR8dMIviVRf/4g1v+yYIACnXl+YA1mDpDmSXi3BAMEKu7KiNhI7iJLmYYg2qIoDJZ4VBxXr624aUbtsmbADjh1yf2Kz3wSls6h0vWX7d6c+02PT9eIpRFlDxBq6pDoFhFLMV7OatFEDpWHa6rVCIQfh14n1F4L0H1WMtsoVdxnb/iDDrsB/BDAZtJpIAUxTv4Kjx+s/qRIcPV5lNgzTvyU8mMOtxakddqdmAytUj4Qv0hnd3u00q5IsEtiL1hyx1vspt2a8XNWjI8VWk/x1GfuG6LQ8ollI+2pBlTI/V3ibOOERmlwh4hZw/jiyrWfHDsH4gSqR+LXjs4rjXouZzx90L1whYmClcLpHTFWlvFxah4TMyj4o+MLgJRvr5cCYreDg2L/HzG4HCUv97PwmFP2G9HGFFwtGoMtuGSJ6tuTQYIlGK1M2hloQG0Qxvj/XhE9g5KqUszbmGJd4ZrFwAZAWJYhX90RRDkWv1rOfw5JmBKUr4kSO/0oAAQViN6bDhiYA7lxkTzmzIATpq7a1IkloSDhhA9ay/aiUI3BAdwcKnyBxTbQs6f+qiSzb/sqsLyphQKH0G7x63JBu1vCHs1dHBgfJkNheDqzpzaPVaHzLhwH4FaQQRbyyB6ITqupLvUhZDQjIpqGwWi33qf+3qWntkG1AZn70mJMyDwHMyff8MIJKlg+fCnK05sFK3ypC7JFIKbBvRBMMCkkPVcy1wXhzyY/zzEQhfMLSL9kDMwk3Fq/KU92zRSVTcWqUMpwOcfH2HO9Gf/94Bw6eqWCSBAfXCita7bPMYTJIFSIEqvA9j2nxGnRlA3GyU0g01tmX7rBMcalH5a5/M4VJnekKMfourJaxPpdGXPpwxyLdEyGcehkeIMq67oeCzFiNICeHgrbsahnB4UrjJZm3GI24URlMqBQUaZMI6e4aOKfTRn2Rt+wkJR3MCHs4GVpNwsYTBMCauWDdZ4W+1Dl/xT1rTqdM9hqFKRvVUblotpK7MOtPVXzrqdLtcXdUJ7WUWuk017vftq6UqAUVQor9ziGhHuzjgH/VKuqowRUiKh95NShm2cDu5/HNouV4UCPJ5Fs8BeKLtU5Rgjb/QWT+bT+Qf3tt2sk1UadHxmFDI7ZZjTvZD+HFXR1wRwryylMMfLPpdvEVWq0yylOOxXK51MIAwzBybmqTv8FuUOuUh+JaC09ZUR+mc3xzgQoObar5VIAG89tMZ9GIifxrkQKeO/GS7aS3O58Fs8PSIpKrLylFoFnAyHoKKJbBUBRD6qYEW9x1uqa4L785FqIqojkIsNViZFIPk+kd7K5sQK+H9CFfYALFKZnV4taPYGFH64mDNl/4KWrhsKhoA1bv073/8Erhm0elJTXGT1kZ73W4p7hkUha7txCvQ5bs1mytCMey4y8O1Cr1QStJfo58RmXlbBfeRv1RzuL0elhPfil9EMEiMXezk3nKFbXtuyF9z8rUhVTZKNnmvObX9Dx5hAgIlrgh9em0Pz0ySXfIOTMJYD0O+YZTh+w0mhV0J9RtG/DtJ1badcuzqblw7p2Kijeog5ij68o5qz5SmTSP04va5nNp9kDIrmQ7jeqzj95c85Z8p8UI5W3Bmf8kpF1kFJvaAmL7aaIDvS9sfou/0p0ax55Oa37YKYaihhHFWdk7ygINMYsnDD7BwBGSday/wvtucgM8UOwA/CG4tebg+HeV+pg/sEw9pOp5IpcTjuZq8qRiqZ7hC9IyHn4AeU/ACdUh1cWhuCA/VaoEat/Qxx9zTq4NDg9JlK/qmqqhE9Cdw4v3RP8N0woZ2Jo5ciSZ3mzCjOHoMcWN4kQPz779x5p9MYFuWey6ZJQzZ0mhCym3y7/GDPEab96pIfNdCr7Uz/GkWkDA06iNhv1jub/zgXuYu3lpfn9PGh/GG//r9+YgJ1CPHthOjjrHrHKKVWhZqUEUv2F/Z6yuNAO45LrI61zNyd77niqGpTUdibsADkGNwKZRrpsOjqJbS1EzmfCLeCKhXjcjzBO2lXqiTiHs1XfoP91tg0htdGBeYDv9S0XqkOgE6t5hfsHUtD+UTdMsbAd0J35cH8deIFNtczpK1N2lumtRp9y71tA7h+yNgxMvLHybEXoycoclDe6DUJffOnOC1K5l3vecasoIM+GSu/1dzXzl0noWo7i99AwTVX15PgidvRC7n3MxciwOIdYxrvyYpfZm4GjLNGIogIQHfh/LhZKV1iaRrIBSNntxMlpT8O19VHKTl9SkNUze34RPQqgHsEZ4USr1GbiT1FOhynIwGxXD2tRaTavHGNK3C1lzUMCzWB/J/JjAXQ5COL9/+1fATn597su9Y+fVUG8/h+UjaLzSdwDGFkI+7yEyWsejojPpiBjrDb3/2T3smDLWChO2cozRarqsabuOh4ljLfakVDtCJWo6KuVeQ7vi3J+FSvGwzUl5/k0TzFcmySbEN0Era8Hh1FZybHgP0nQzYiJ8ay7DgYZvYmlmBeoCsPCL+DV7l1Ayj1f96aBNgBZglb0ceiosg2/OG90GHXsn9x6RhKy0PbvG0Xt5ZSnVDEDze4TDjB2pPMjrXimaERGlvSJ/u2UlMSj28706HEdek6tmYmxgAvNqeDcywz9v3vAPdpMNTX52KyyFxiZ0YJs/ms1O3pxlIQDKAPNMkr3O6XhidlX7WJxc9zTPcFUDyPXqO5TPhTJv6LGf1LSensKrQOw6kLkjU4ndVDdRxPk/OVbHhIqUQQOi/4RPlK+QWyGo27Q82t8RkCoaxHkTInFxtov6v+Z94jLTpG33J1SM/0zwzUBKh8ohlLZELx3tJL7p8Rynp92vKwGXKYOqjMK7heraEfq4pj4xnNgJ6B+M/lR7xOqZaaUPGFYKJFNGUxZNwhpN+Cl20JtReWFjCtt/iPGrgEznuVXp4zdsIxANGt1Y6Wjso031N8VXx36JEdQOL2JDb4aIlCRU8sf04B65YT9dNPJUKFkTP81d4kpR/LVYn4yysHv4SuuLbUFbNiyWXaxaMWExy0072poCgiwvlLqKICTDMnUfRGSjOh8YuHla7LaGROY7M++MrRRdb0D7cooS1jpCNtvZgjnwgZTYWk77QouBTSBsyx6ObS6B2cnmmQnUZT8LlYW80iGalP5rYdVzUPKmZc+DOFTdXo7w8bG6v0wohqHY7FiL/UnWwFtF112/jrA0AwyOdsiw4iTTvNQHeT5XS6gAEGri9yB6FGaNOmZo5XK+cQY1GTdG+1ncvtuvDNANiuawfG/pc0iMI8yPuGXeLat4xHLcfj4uacj+vvbqSdi5UTUO0JGJFd8kg9eCi/c7L8OSVSI6eLuRRbw4p6eqi1XY6Zxtlnrvq38z4V1JxpKQhVlQLwOR3Lw4BjpZbx5NCdwZ2gjJ8T2QvFc3g06Ds28In1zNtumJ5m6DRircJpoHF/EFLGLjuy1guXQUHF8GHhWj5xl1HljueAEv6jRY7OAwv8ootBLC938t+Ss56DKs1OpPU5mRHTamlH4zsBeJJq7aa87tsUp+YIJ/LJ8su4Ol6b5AaK5jwg0mXHVp7ZGsdPXbNo0UCkrBZ89yvD1UnXIXiDhahG7bgbbkMlLSjtHhDYFWE+pBv5cg4Lc13Ci8sTwN8trogWHu7NwVHqUnlH/k+a8bp0D+zr7WugDV3oMEUwYT09z3lK9L3u/fvchQ6HRwzMB7NvkYwn6Bm8Gcxeg0wt4FyZgyLbolqv/Q8yB20U/polyISsRQpJWgpUlR4I0RbOyTk+85nvvUMFaVLcrhRHx3TesxGDBfKRFdTlT6zXF8j1zoPiaiTQwMLokabjjbdMJT7/QsV5PjfYsYaQ6bCECF/ommpR+JFBmXlhfJtTqbhYi9htHZyMhIZFjmjsVpL3KY8RXxgM4PxgJ51csulC4a/18raoEd4ah+HEzIAzRQ8SG5P8HH7W6hbx+SGt/KvL5y6Xaqo6odGp6/2BhnOhE99w02XNuxABk84dn2P06auZch0g8Dx8ogSTh2D662fbjr18a83ORSn3gTJ74rqakBOvw6g3ZK+eb/Jz7hM802km+x8TMb/cyWkUghFLnIF6GU19VWStkfKf/q0EVCg4umZ4nyBMG4JHjjEx6CGuf08fM73fNeeZ7jolUhhFuwn+ZRS3sSgmAK8ZueGiWWRnROzfX5djkAI/YfpsQm6iaO89LLWvVfdoO9LHu5vwkAQNPwbDYJ8FloQTcmaOEY9qq9HwdlhRiYaoKMkDjSSbQ4E+pkd4qiHT+lYpH+LdPbgHPjt8jQ+1xYeuPWpCTAr+UOX9CW7N0BL9wNwVsNdFo/G191phQXfhhMN8VXokp8AQ885x9OXzM/KJgd/CqRLmIshqa/lN7bQ6Sg8Fs/ZMb/aOJthsXg5HKyzuiqXgqinZmvMb6wGdG/zcTE+Hi3u3Xc+ZSqCrOmmUc04QqPJD9lGtCwD3RohfXGiLUlUhuVhtE2ECEHePPne3PO1xozj2JtNP6pOW6DCsJ22Qj2bDUOhOAVSYMEYiJhdFBrke1fS8X/3Llo+68C/4YcfpDWZJJKoK/1W2DOnSk+nPWXPBcddpthY9uAicjpMPdu9+xVwuawXhN2nlJV+0pJwWjwnT2PJAe6LeBA6qoDpYZFC5f03CmpEXiwL3Hhb3/UY0CcjQCRUcJLR/fuvVATWiaxAW1quA+82pedDtyXuVxK7EXEx50xW7K2i2oJAUXIhGmynl+cN5VwhDBo9Hoo4wUm2rIuMZ78D5j8ueUWwtEZ4NwtRD7y50eAvfCqy/XNNSbQn8cQK8bpudUxj6sWET3xsAo3CZetms/6BS0HMGsCpAj/f5Fn4XOYaf+oAry8c0BgrvcHRSui2FMdaj+MbRVNfgyWGUppYevlzbBcN2n1ckQSW+F36+FLMXrS4ZMUnskBuwOmO0V0A5jyP/gHHVhJLj55hIvmzw2VCZjvt4FXtD+tN7w2F3oQuA2xMpxC1pE0srY0BOXYUPfAl/rHhbg6zww2AC8sGqzPXhGA+Ui7s4ckMsE/Ds4+pF4GfqQVLmFoCMhfZ3aL+ddHdQGXFiB9NzbMx0QppC9i1fbYuarT8ooTn2F6MwYsLx8xGLG5oz9gpXNpVZ/PUjghfFA805dqrE9ZqUAM0mfUYdJUODlQ1DZyGhqPgPgN0KkrxKaktIQq1TQkM3dGVu3zOVti3Dfl7FvamvZdVUjdg5mdf2SEP/RxXN/wjyp/36Bc7td1daVQeFK3sayy7766Y6UQqmEZGqkw3gsKmU32TDNOfluj5Dl/r9yZ2adpMWIfq6+Bz00Bnw8aWKU3Y3/IH/+ttP2V4Qn7Epm/Jw8hJlmByYNA/RuVWSZMqKZzLDyWJRz0lcZWa9cqAHtNVUgSH1TsYXXbz2fUutDGVJo917ic2kJrXVYV1Gr9wbVDw0XBR0jbJpNIpUziMN44EpPmWeL0foj6DY5A4jPl+E8aARJtHGcYc7HggUGGXMs/2zPSYRgeC2LT4SLs3RkczbaZYkQmybFIBdeExndagh3JOymF4EAS2zmtbpNePEDv0sjMQjnubnZWNyQbTxvBi0OiL3DlgbhcbcAOfs7GGQpi/H5UGfsGka6BdFHf4PgowU2hMwqr/d0ItG86kmee0LMS3KYaxxTrAklMswtHGEpW/RJE3XjQVWJJSFpZbpg1bWW2nlVS0ybSAtOeCAbe6OccIU4OqygSPSjttQdL4O9AQIpE6Z8DlAzZFir/ibogcuW7J9198jztScw+PS3ylwRzgZMc1+BJweBN2t5uq4EQyQaVev/QsOTXy4BWiycSubAEPTf1WwyZiw9wJLAdDKiBvdgPNWaY4+7rPWWAlTPrvQqe+RmA7iWazBLhpWZ55PtJDrvnlWpRvsy4VNuxomZUp/Q5hfHEg2AEEjLe+8QjKIeTfQNtPa5WOeFvdIeKcTeoCQKAMURb2G5mTpBpWkswzc2AbHyb94VSSUvFKn+g+380Tk8txzIFpJKonCXluO6+EX3n2wMlLMfw3xRNufaNjxFUBfgHhdDyqZSvt36TewTItV9vSYSNGOI0yWG+2z+MzVeZVPPWAHFmy7SqUMhdrrRxwDI17eyQLPWbQ2xwU1otsFLdQOpAtgBE+EQJgeM2Ll8FgqrZxNGLMXvee2U+QRUBVQGYy+54VU65Qwt424aHgK2xJVMJvEeBUp3+YAST401mID/8uasIX0Cms7gePvKe4cFl6OPunrbO5PU10u9x2eo0Zs1UIl+rXbaEDBTO4BJVmx1sit/fHFsBPnqnCjJxKfWvHUqF5j6nCJjadZOE0wtnZJxDUL11TgdZwbSjH0ahh+R03qp7uzP0uoI60cyycFlY77JfSSFR0ixPBW3AbEqic/TYj4iTzRyZunA9yjXRFU7bra/Hr+iLVL768rbnlTM+cSxctrDq3wqFJw42RrBY00BmtsClOUqz4UFf/uviPxFE1etRvLdfZIiwCU84EhDn6qOABIKA+TmVTckAAva59ZnnPtPHmMX2fIgu0jKGwk4jpkCLNyFzdq8Il8F7ITFU/pMVQCYVz6GwQ49Ov5ALvqyAcR/7gGzI1uX19RgtPIw0kfqz3o0Gd0FA4a4Cuzjk5WAQPGBtG+QjwIxC55yXFQK5LYQ7LyxUnz93JKqofoWj+VH5RYqg0bmezX/2T+ELnsvHKDNZzX4nbmTj1qAUZqwyN7GKBNgA6qRBwmlM1U4rHq9WJyS4rT71dAvD1XxCSFOX6tw8T/JoxVZxRAN6AoDmdbGKroWUwiNFJ+7xfJrCXghRRsSqAhRZFGz90NFfiVMHXluTz0rq1YPRRkgxY9IPY0J9FoUs/NXitw8aoseSD2oV3/yn5WEDu7kLt2Uz+uuP1a06hQlKZL/Z5mxP2sjFrfcBsI5VUUuoiJA7357G+O0H+rZqLRfbXGnrv5y2QfbPkTend1etKXTvUWHe5h1TRh5/3VhB8qcI7cf5B4j7voOuLHBT837/PVaNHsgbikmNIvQ2/Wiy958ghGPL7IVhe5n8EaEoXFZcgCRInNoWPotM8dGI0yihAcEbobMf4B1VMKZPX2ZNDrq0GjZ0wEgpsdl7Ufx7z6rpRDz6a5GmJXW5tz6OAEv55vUv/HoGkSECaTySbbMDzxdfJit7DWJwwjoR5cbQFiNfbaczf2woSznMJTl64naiVh+54zWWt2Q6C3nazE+5rRwLY1v1YMk+8CloAxIxeZiIm6SePJXazwYAvEOF7JoCDlhfkwM6L6LMDcL9fb58m6ukmqaoy5pHpQxu7zpwbXjtreCgLS4rq5/pMYDQ8fIRklRNhPod4CORQJTLpcRThauuJB9y5C9bMJDZsX75NUWLrgPgiYBl+OZCOSa7Qn2zPKkgdB5nqkhSGUvhAEsfstaYajnGfxxEEYGIrftP1bNSDqSsu3VA8Ze0ILlSnyrSCeIQk73DAA5hPSvx4uP5bUdzaMB9/PXpfBgOSQ8ZkJVehsdrpBTAaU5bv4zsAECMtMS6IsBNmw1TjAPMmhiNVfr37CFrWDkDEFtKel0izCqmRwTo41WnXIMre9/MnQTo0Zn9Ev6Q4rsPk+7uAYSabYHCfgIu046LjdYknq/jzVqhumYfuWOReljShzJ/te9MDuv/HMz30sA3OC0j2k0QSDp7QV2xcBfNrABNoVtGJTgqY2zD0TuIL4CW3PJSLoaCsX77/X+udGvP72g9zjeEbxLKGAeY0+0p7w4HruB13j6sAZnerMCBHlo41VuFJtLFq60oumCwbW,iv:zsu1VvaRVbzp/mw4BqZSOyk4GwIkSlV133j7pPpKfpo=,tag:jnJJoIWCpgVcIwq7/a+wzw==,type:str]",
-	"sops": {
-		"kms": null,
-		"gcp_kms": [
-			{
-				"resource_id": "projects/two-eye-two-see/locations/global/keyRings/sops-keys/cryptoKeys/similar-hubs",
-				"created_at": "2024-01-30T21:02:48Z",
-				"enc": "CiUA4OM7eCxHPiBqy6mOWR2S3PJbLSWYZXzFOQnZJ8lrTylqCTNlEkgAjTWv+o7jO3nOhYDxVi96kSIKuPF6qw/LbnuwkQhW1Q2KPYvEqWxnuSlGmdbgaB3ue6zs963VkoviKthuemqMELyCzwFLdR8="
-			}
-		],
-		"azure_kv": null,
-		"hc_vault": null,
-		"age": null,
-		"lastmodified": "2024-04-22T16:17:30Z",
-		"mac": "ENC[AES256_GCM,data:gffE9sWu1rpxinlkzjpr8ymOQHTorrac1dkk6dUVfPd2uV88KEw5ayG50VKl8+onKIv3kDjcf0aO2yt7SochAMCK8d/wLO8HUkuJQwmReJnWB0UmjzJwmp4OZ9/boxW4N1lzE8ZCN+Go4bIqZG+FjUF7kcby4XCkKQuPqzNG8fM=,iv:fNNHSyd0ejzyTm5G6tB7eaB0WFdRug5DW9EWY9Wp/CM=,tag:R0+JlMJHPTLgmmCUjBvcLQ==,type:str]",
-		"pgp": null,
-		"unencrypted_suffix": "_unencrypted",
-		"version": "3.8.1"
-	}
-}
